requirements.txt
streamlit
pandas
python-docx==1.1.0